misc sample fixes
