+++ sdk/formrecognizer/azure-ai-formrecognizer/samples/v3.2-beta/sample_analyze_general_document.py
@@ -7,18 +7,14 @@
 # --------------------------------------------------------------------------
 
 """
-FILE: sample_analyze_document_async.py
+FILE: sample_analyze_document.py
 
 DESCRIPTION:
     This sample demonstrates how to extract general document information from a document
     given through a file.
 
-    Note that selection marks returned from begin_analyze_document() do not return the text associated with
-    the checkbox. For the API to return this information, build a custom model to analyze the checkbox and its text.
-    See sample_build_model_async.py for more information.
-
 USAGE:
-    python sample_analyze_document_async.py
+    python sample_analyze_document.py
 
     Set the environment variables with your own values before running the sample:
     1) AZURE_FORM_RECOGNIZER_ENDPOINT - the endpoint to your Cognitive Services resource.
@@ -26,7 +22,6 @@ USAGE:
 """
 
 import os
-import asyncio
 
 def format_bounding_region(bounding_regions):
     if not bounding_regions:
@@ -39,11 +34,10 @@ def format_bounding_box(bounding_box):
     return ", ".join(["[{}, {}]".format(p.x, p.y) for p in bounding_box])
 
 
-async def analyze_document():
+def analyze_document():
     path_to_sample_documents = os.path.abspath(
         os.path.join(
             os.path.abspath(__file__),
-            "..",
             "..",
             "..",
             "./sample_forms/forms/form_selection_mark.png",
@@ -51,7 +45,7 @@ async def analyze_document():
     )
     # [START analyze_document]
     from azure.core.credentials import AzureKeyCredential
-    from azure.ai.formrecognizer.aio import DocumentAnalysisClient
+    from azure.ai.formrecognizer import DocumentAnalysisClient
 
     endpoint = os.environ["AZURE_FORM_RECOGNIZER_ENDPOINT"]
     key = os.environ["AZURE_FORM_RECOGNIZER_KEY"]
@@ -59,13 +53,11 @@ async def analyze_document():
     document_analysis_client = DocumentAnalysisClient(
         endpoint=endpoint, credential=AzureKeyCredential(key)
     )
-
-    async with document_analysis_client:
-        with open(path_to_sample_documents, "rb") as f:
-            poller = await document_analysis_client.begin_analyze_document(
-                "prebuilt-document", document=f
-            )
-        result = await poller.result()
+    with open(path_to_sample_documents, "rb") as f:
+        poller = document_analysis_client.begin_analyze_document(
+            "prebuilt-document", document=f
+        )
+    result = poller.result()
 
     for idx, style in enumerate(result.styles):
         print(
@@ -74,8 +66,8 @@ async def analyze_document():
             )
         )
 
-    for idx, page in enumerate(result.pages):
-        print("----Analyzing document from page #{}----".format(idx + 1))
+    for page in result.pages:
+        print("----Analyzing document from page #{}----".format(page.page_number))
         print(
             "Page has width: {} and height: {}, measured with unit: {}".format(
                 page.width, page.height, page.unit
@@ -84,7 +76,7 @@ async def analyze_document():
 
         for line_idx, line in enumerate(page.lines):
             print(
-                "Line # {} has text content '{}' within bounding box '{}'".format(
+                "...Line # {} has text content '{}' within bounding box '{}'".format(
                     line_idx,
                     line.content,
                     format_bounding_box(line.bounding_box),
@@ -100,7 +92,7 @@ async def analyze_document():
 
         for selection_mark in page.selection_marks:
             print(
-                "Selection mark is '{}' within bounding box '{}' and has a confidence of {}".format(
+                "...Selection mark is '{}' within bounding box '{}' and has a confidence of {}".format(
                     selection_mark.state,
                     format_bounding_box(selection_mark.bounding_box),
                     selection_mark.confidence,
@@ -123,7 +115,7 @@ async def analyze_document():
             )
         for cell in table.cells:
             print(
-                "...Cell[{}][{}] has text '{}'".format(
+                "...Cell[{}][{}] has content '{}'".format(
                     cell.row_index,
                     cell.column_index,
                     cell.content,
@@ -138,14 +130,14 @@ async def analyze_document():
                 )
 
     print("----Entities found in document----")
-    for idx, entity in enumerate(result.entities):
+    for entity in result.entities:
         print("Entity of category '{}' with sub-category '{}'".format(entity.category, entity.sub_category))
         print("...has content '{}'".format(entity.content))
         print("...within '{}' bounding regions".format(format_bounding_region(entity.bounding_regions)))
         print("...with confidence {}".format(entity.confidence))
 
     print("----Key-value pairs found in document----")
-    for idx, kv_pair in enumerate(result.key_value_pairs):
+    for kv_pair in result.key_value_pairs:
         if kv_pair.key:
             print(
                     "Key '{}' found within '{}' bounding regions".format(
@@ -155,7 +147,7 @@ async def analyze_document():
                 )
         if kv_pair.value:
             print(
-                    "Value '{}' found within '{}' bounding regions".format(
+                    "\nValue '{}' found within '{}' bounding regions".format(
                         kv_pair.value.content,
                         format_bounding_region(kv_pair.value.bounding_regions),
                     )
@@ -165,9 +157,5 @@ async def analyze_document():
     # [END analyze_document]
 
 
-async def main():
-    await analyze_document()
-
 if __name__ == "__main__":
-    loop = asyncio.get_event_loop()
-    loop.run_until_complete(main())
+    analyze_document()
